refactor(server): move types into separate files
